Changed the index of Nodes/Edges to private with getters & Setters
Refactored the index of nodex/edges to private with getIndex() and setIndex()
Extracted the Index and its functions to VItem, because both node and edge
have an index. Added a Warning to SetIndex because there are seldom situations,
in which you want to set the index but you really should handle that with care.