Big Code Cleanup #3
Recognized further Methods unneeded and ugly in VGraph/MGraph optimized that
refactored man of the VGraph methods to fit an unified Signature

- Each add-Method now takes the two parameters of its specific V- and MElement to be added
- VGraph.addSubSet optimized
- VGraph.ControlPointinRange adapted to fit EdgeinRange
- Removed some Methods (e.g. VGraph.getNodeName(int i) replaced by VGraph.getMGraph.getNode(i).name)

Found a Bug that would result in wrong SubSet-Names when loding an GravelML Graph with SubSets.
Because this is one of the central classes - several other classes where affected and (mainly automatically by eclipse) refactored. I hope i didn't produce Bugs.