Divided the ShapeMOdification into a Subclass of VHyperGraphic, and named it VHyperShapeGraphic
and all its Mouse-Drag-Actions into a Subclass of DragMouseHandler-Interface,
called DragShapeMousehandler.
CircleDragListener is the first one of these creating a Circle with
radius of Drag Distance and Origin of the Beginning of the Drag.

This is all included into CHyperEdgeShapeParameters and the Circle can be
created and its Values reach the Parameters-Class and they are printed.
Next Step is to Set them in Parameter-Fields
	(that might be hidden and only seen in experts mode)