The error is corrected, subcurve selection WORKS
It was:
elevating multiplicity of u1 (near to the end, that ist u1>Knots.get(max-2*degree)), so somewhere in the last degree Polynomials affects also ControlPoints/Knots of the start-Stuff

Now u1 is elevated, start-area updated (if u1 too near) and after that u2 elevated and then the curve cut.

That results in a subcurve beginning in u1 following the curve (if u1>u2 and curve unclamped&closed thereby rolling over start/end) ending in u2. If curve is defined clockwise, thispart is clockwise, otherwise anticlockwise.